直接查包裹ok了

+++ src/main/java/controller/userController/OrderDetailController.java
@@ -30,7 +30,7 @@ public class OrderDetailController implements Initializable {
         System.out.println("加载订单详情界面");
     }
 
-    public void init(int order_id,String sendAdd,String receiveAdd){
+    public void init(long order_id,String sendAdd,String receiveAdd){
         // "1157843878343"、"广州"、"阳江"这三个字段需要数据填充
         orderId.setText("运单编号:"+order_id);
         send2ReceivePath.setText(sendAdd.split(";")[1]+"->"+receiveAdd.split(";")[1]);
